(fix) Correct parsing of EBNF comments
Originally, `(*)` would be treated as a comment, which is not correct
EBNF syntax. `(*)` would be treated as a group, containing the symbol
"*", which would be a syntax error entirely. This commit changes the
behavior of parsebnf's EBNF parser to not parse `(*)` as a comment.

Examples of correct EBNF comments are as follows:

* `(**)`
* `(***)`
* `(* *)`
* `(*A comment containing some words*)`
* `(* A comment containing some words and spaces outside them *)`
* `(*   Yet another comment, weird leading/trailing spaces*)`